+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### NEW
 
 * _Linked_ project to `S7Central`.
+* _Created_ function to parse versioning information. -- Credits to LaughingLeader 💯
 ## [1.1.0.0] --- 8th November 2020 --- **_Scriber_**
 
 ### NEW